rename package name to dql
